Move config parsing to separate folder
