+++ src/main.js
@@ -1,12 +1,9 @@
 #!/usr/bin/env node
-
 import fs from "fs";
-import path from "path";
-import os from "os";
-import toml from "toml";
 import program from "./commander/setup.js";
 import { getOptions } from "./commander/argHandlers.js";
 import generateCompletion from "./openai/generateCompletion.js";
+import { readConfigFile } from "./utils/readConfig.js";
 
 program.action(async (files) => {
   try {
@@ -60,24 +57,6 @@ program.action(async (files) => {
     process.exit(1);
   }
 });
-
-function readConfigFile() {
-  
-  const homeDir = os.homedir();
-  const configFilePath = path.join(homeDir, "./genereadme-config.toml");
-
-  if (fs.existsSync(configFilePath)) {
-    try {
-      const configFileContent = fs.readFileSync(configFilePath, "utf-8");
-     
-      return toml.parse(configFileContent);
-    } catch (error) {
-      console.error("Error parsing the config file:", error.message);
-      process.exit(1);
-    }
-  }
-  return {};
-}
 
 function main() {
   program.parse(process.argv);
